automation/requirements.txt
requests
python-dotenv
aiohttp
